backend/giro/requirements.txt
djangorestframework==3.3.2
markdown==2.6.5
django-filter==0.11.0
